requirements.txt
streamlit
opencv-python
pillow
numpy
pandas
python-doctr[torch]
pytesseract
google-generativeai
